LF-4089 Extract MoreComponent into own file; add hover state
