+++ packages/webapp/src/components/TileDashboard/MoreComponent/styles.module.scss
@@ -1,0 +1,107 @@
+/*
+ *  Copyright 2024 LiteFarm.org
+ *  This file is part of LiteFarm.
+ *
+ *  LiteFarm is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  LiteFarm is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
+ */
+
+.moreContainer {
+  position: relative;
+  overflow-x: visible;
+}
+
+.moreButton.moreButton {
+  // increase CSS specificity beyond textButton class
+  display: flex;
+  width: 100%;
+  height: 40px;
+  min-width: 90px;
+  padding: 0px 4px;
+  justify-content: center;
+  align-items: center;
+  gap: 4px;
+
+  color: var(--Btn-primary-pristine, #ffcf54);
+  font-size: 12px;
+  font-weight: 400;
+  letter-spacing: -0.264px;
+
+  border-radius: 4px;
+  background: var(--Colors-Primary-Primary-teal-700, #266f68);
+}
+
+.moreContent {
+  position: absolute;
+  top: 48px;
+  right: 0;
+
+  display: grid;
+  grid-template-columns: repeat(2, max-content);
+  grid-auto-flow: row;
+
+  gap: 8px;
+  padding: 8px 4px;
+
+  border-radius: 4px;
+  background: var(--Colors-Primary-Primary-teal-50, #ebf5f4);
+  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
+}
+
+.contentItem:only-child {
+  margin-right: -8px;
+}
+
+.contentItem {
+  width: 100%;
+  min-width: 78px;
+  max-width: 90px;
+}
+
+.moreButton:hover {
+  background: var(--Colors-Secondary-Secondary-green-700, #558f70);
+
+  color: var(--Colors-Accent-Accent-yellow-50, #fff8e6);
+
+  svg path {
+    stroke: var(--Colors-Accent-Accent-yellow-50, #fff8e6);
+  }
+}
+
+.moreButton.open {
+  background: var(--Btn-primary-hover, #e8a700);
+
+  color: var(--Colors-Accent---singles-Green-dark, #0c3611);
+
+  svg path {
+    stroke: var(--Colors-Accent---singles-Green-dark, #426f57);
+  }
+}
+
+.moreButton.selected {
+  background: var(--Btn-primary-pristine, #ffcf54);
+  color: var(--Colors-Secondary-Secondary-green-800, #426f57);
+
+  svg path {
+    stroke: var(--Colors-Secondary-Secondary-green-800, #426f57);
+  }
+}
+
+.moreButton.selected:after {
+  content: '';
+  width: 8px;
+  height: 8px;
+  border-radius: 50%;
+  background-color: var(--Colors-Secondary-Secondary-green-900, #325442);
+
+  position: absolute;
+  top: 4px;
+  left: 4px;
+}
