+++ packages/webapp/src/components/TileDashboard/styles.module.scss
@@ -31,7 +31,7 @@
   font-size: 16px;
   font-style: normal;
   font-weight: 400;
-  line-height: 24px; /* 150% */
+  line-height: 24px;
   text-transform: capitalize;
   height: 24px;
   margin-bottom: 8px;
@@ -50,8 +50,6 @@
   align-items: center;
   gap: 8px;
   flex-shrink: 0;
-
-  // outline: 1px solid white;
 }
 
 .gridContainer {
@@ -61,8 +59,6 @@
   flex-flow: row wrap;
   align-content: center;
   gap: 4px;
-
-  // outline: 1px solid paleturquoise;
 }
 
 .gridItem {
@@ -78,92 +74,4 @@
 
 .gridItem.fixed {
   flex: 0 0 auto;
-}
-
-.gridItem.hidden {
-  visibility: hidden;
-}
-
-// More component
-.moreContainer {
-  position: relative;
-  overflow-x: visible;
-}
-
-.moreButton.moreButton {
-  // increase CSS specificity beyond textButton class
-  display: flex;
-  width: 100%;
-  height: 40px;
-  min-width: 90px;
-  padding: 0px 4px;
-  justify-content: center;
-  align-items: center;
-  gap: 4px;
-
-  color: var(--Btn-primary-pristine, #ffcf54);
-  font-size: 12px;
-  font-weight: 400;
-  letter-spacing: -0.264px;
-
-  border-radius: 4px;
-  background: var(--Colors-Primary-Primary-teal-700, #266f68);
-}
-
-.moreContent {
-  position: absolute;
-  top: 48px;
-  right: 0;
-
-  display: grid;
-  grid-template-columns: repeat(2, max-content);
-  grid-auto-flow: row;
-
-  gap: 8px;
-  padding: 8px 4px;
-
-  border-radius: 4px;
-  background: var(--Colors-Primary-Primary-teal-50, #ebf5f4);
-  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-}
-
-.contentItem:only-child {
-  margin-right: -8px;
-}
-
-.contentItem {
-  width: 100%;
-  min-width: 78px;
-  max-width: 90px;
-}
-
-.moreButton.open {
-  background: var(--Btn-primary-hover, #e8a700);
-
-  color: var(--Colors-Accent---singles-Green-dark, #0c3611);
-
-  svg path {
-    stroke: var(--Colors-Accent---singles-Green-dark, #426f57);
-  }
-}
-
-.moreButton.selected {
-  background: var(--Btn-primary-pristine, #ffcf54);
-  color: var(--Colors-Secondary-Secondary-green-800, #426f57);
-
-  svg path {
-    stroke: var(--Colors-Secondary-Secondary-green-800, #426f57);
-  }
-}
-
-.moreButton.selected:after {
-  content: '';
-  width: 8px;
-  height: 8px;
-  border-radius: 50%;
-  background-color: var(--Colors-Secondary-Secondary-green-900, #325442);
-
-  position: absolute;
-  top: 4px;
-  left: 4px;
 }
